Represent search query in the url of overview #64
