+++ package.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "classnames": "^2.2.5",
     "docuri": "^4.2.1",
+    "history": "^4.6.1",
     "lodash": "^4.17.4",
     "page-metadata-parser": "git://github.com/Treora/page-metadata-parser.git#npm-testable",
     "pouchdb-browser": "^6.1.0",
@@ -26,6 +27,7 @@
     "react-dom": "^15.4.1",
     "react-redux": "^5.0.1",
     "react-router": "^3.0.2",
+    "react-router-dom": "^4.0.0",
     "readability": "git://github.com/mozilla/readability.git#2b28d60",
     "redux": "^3.6.0",
     "redux-act": "^1.1.0",
