+++ src/overview/components/Overview.jsx
@@ -1,8 +1,6 @@
 import React from 'react'
-import { connect } from 'react-redux'
+import createBrowserHistory from 'history/createBrowserHistory'
 
-import * as actions from '../actions'
-import { ourState } from '../selectors'
 import ResultList from './ResultList'
 
 import styles from './Overview.css'
@@ -12,7 +10,7 @@ class Overview extends React.Component {
         return <div>
             <input
                 className={styles.query}
-                onChange={e=>this.props.onInputChanged(e.target.value)}
+                onChange={e=>this.handleInputChange(e.target.value)}
                 placeholder="Search your memory"
                 value={this.props.query}
                 ref='inputQuery'
@@ -27,17 +25,18 @@ class Overview extends React.Component {
             this.refs['inputQuery'].focus()
         }
     }
+
+    componentWillMount() {
+        const query = new URLSearchParams(this.props.location.search).get('q')
+        if(query) {
+            this.props.onInputChanged(query)
+        }
+    }
+    handleInputChange(query) {
+        const history = createBrowserHistory()
+        history.push({ pathname: '/overview/overview.html',search: `?q=${query}` })
+        this.props.onInputChanged(query)
+    }
 }
 
-const mapStateToProps = (state) => ({
-    query: ourState(state).query,
-    searchResult: ourState(state).searchResult,
-})
-
-const mapDispatchToProps = (dispatch) => ({
-    onInputChanged: input => {
-        dispatch(actions.setQuery({query: input}))
-    },
-})
-
-export default connect(mapStateToProps, mapDispatchToProps)(Overview)
+export default Overview
