+++ src/overview/index.js
@@ -2,5 +2,6 @@ import reducer from './reducer'
 import * as actions from './actions'
 import * as epics from './epics'
 import Overview from './components/Overview'
+import Main from './components/Main'
 
-export default {reducer, actions, epics, components: { Overview }}
+export default {reducer, actions, epics, components: { Overview, Main }}
